// frontend/src/api/generated.d.ts
export interface components {
    schemas: {
        AddFeedRequest: {
            url: string;
        };
        Article: {
            id: string;
            feedId: string;
            guid: string;
            title: string;
            url: string;
            isRead: boolean;
            feed: components["schemas"]["ArticleFeed"];
        };
        ArticleConnection: {
            articles: components["schemas"]["Article"][];
            pageInfo: components["schemas"]["PageInfo"];
        };
        ArticleFeed: {
            id: string;
            url: string;
            title: string;
            isSubscribed: boolean;
        };
        ErrorResponse: {
            message: string;
        };
        Feed: {
            id: string;
            url: string;
            title: string;
            fetchedAt: string;
            isSubscribed: boolean;
            /** Format: int32 */
            unreadCount: number;
        };
        LoginRequest: {
            username: string;
            password: string;
        };
        LoginResponse: {
            user: components["schemas"]["User"];
        };
        PageInfo: {
            hasNextPage: boolean;
            endCursor?: string;
        };
        User: {
            id: string;
            username: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}